src/Animal.java
import com.sun.org.glassfish.gmbal.Description;
import com.sun.tools.javac.util.Name;

/**
 * Created by cmitchell on 3/29/17.
 */
public class Animal {
    private String name;
    private String species;
    private String breed;
    private String description;


    public Animal(String name, String species, String breed, String description) {
        this.name = name;
        this.species = species;
        this.breed = breed;
        this.description = description;

    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSpecies() {
        return species;
    }

    public void setSpecies(String species) {
        this.species = species;
    }

    public String getBreed() {
        return breed;
    }

    public void setBreed(String breed) {
        this.breed = breed;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }


}
